improve swiper initialization. Sepration of concerns
