+++ widgets/carousel/widget.settings.ts
@@ -1,10 +1,3 @@
-import {
-  onTileExpand,
-  initializeInlineSwiperListeners,
-  onTileClosed,
-  onTileRendered,
-  onExpandedTileCrossSellersRendered
-} from "./widget.extensions"
 import { registerExpandedTileCrossSellersRendered, registerWidgetInitComplete } from "widgets/libs/tile.listeners"
 import {
   addAutoAddTileFeature,
@@ -16,6 +9,13 @@ import { addCSSVariablesToPlacement } from "widgets/libs/widget.layout"
 import getCSSVariables from "widgets/libs/css-variables"
 import { Sdk } from "@stackla/ugc-widgets"
 import { getConfig } from "./widget.config"
+import { initializeInlineSwiperListeners } from "./components/inline-swiper.loader"
+import {
+  onTileClosed,
+  onTileExpand,
+  onTileRendered
+} from "@libs/components/expanded-tile-swiper/expanded-swiper.loader"
+import { onExpandedTileCrossSellersRendered } from "@libs/components/expanded-tile-swiper/product-recs-swiper.loader"
 
 declare const sdk: Sdk
