+++ widgets/libs/components/expanded-tile-swiper/product-recs-swiper.loader.ts
@@ -1,0 +1,28 @@
+import { initializeSwiper } from "@libs/extensions/swiper/swiper.extension"
+
+export function onExpandedTileCrossSellersRendered(tileId: string, target: HTMLElement) {
+  // initialize swiper for cross-sell products
+  if (target && target.shadowRoot) {
+    const swiperCrossSell = target.shadowRoot.querySelector<HTMLElement>(".swiper-expanded-product-recs")
+
+    if (swiperCrossSell) {
+      initializeSwiper({
+        id: `expanded-product-recs-${tileId}`,
+        mode: "expanded-product-recs",
+        widgetSelector: swiperCrossSell,
+        prevButton: "swiper-exp-product-recs-button-prev",
+        nextButton: "swiper-exp-product-recs-button-next",
+        paramsOverrides: {
+          grabCursor: true,
+          allowTouchMove: true,
+          loop: true,
+          on: {
+            beforeInit: swiper => {
+              swiper.slideToLoop(0, 0, false)
+            }
+          }
+        }
+      })
+    }
+  }
+}
